requirements.txt
dash
dash-bootstrap-components
databricks-sdk
python-dotenv
networkx
matplotlib
dash-cytoscape
PyGithub
chromadb>=0.4.0
sentence-transformers>=2.2.0